+++ src/xtd.drawing.native.fltk/src/xtd/drawing/native/fltk/font_family.cpp
@@ -34,12 +34,12 @@ namespace {
   }
 }
 
-intptr_t font_family::create(const std::string& name) {
-  if (name == ".AppleSystemUIFont") return reinterpret_cast<intptr_t>(new std::string(name));
-  return  families().find(name) != families().end() ? reinterpret_cast<intptr_t>(new std::string(name)) : 0;
+intmax_t font_family::create(const std::string& name) {
+  if (name == ".AppleSystemUIFont") return reinterpret_cast<intmax_t>(new std::string(name));
+  return  families().find(name) != families().end() ? reinterpret_cast<intmax_t>(new std::string(name)) : 0;
 }
 
-void font_family::destroy(intptr_t font_family) {
+void font_family::destroy(intmax_t font_family) {
   delete reinterpret_cast<std::string*>(font_family);
 }
 
@@ -59,26 +59,26 @@ std::vector<std::string> font_family::installed_font_families() {
   return std::vector<std::string>(families().begin(), families().end());
 }
 
-int32_t font_family::get_cell_ascent(intptr_t font_family, int32_t em_height, bool bold, bool italic, bool underline, bool strikeout) {
+int_least32_t font_family::get_cell_ascent(intmax_t font_family, int_least32_t em_height, bool bold, bool italic, bool underline, bool strikeout) {
   fl_screen_dc screen_dc;
   fl_font font(font_family, em_height);
   return screen_dc.ascent(font.font(), pixel_to_native_font_graphics_untit(font.size()));
 }
 
-int32_t font_family::get_cell_descent(intptr_t font_family, int32_t em_height, bool bold, bool italic, bool underline, bool strikeout) {
+int_least32_t font_family::get_cell_descent(intmax_t font_family, int_least32_t em_height, bool bold, bool italic, bool underline, bool strikeout) {
   fl_screen_dc screen_dc;
   fl_font font(font_family, em_height);
   return screen_dc.descent(font.font(), pixel_to_native_font_graphics_untit(font.size()));
 }
 
-int32_t font_family::get_line_spacing(intptr_t font_family, int32_t em_height, bool bold, bool italic, bool underline, bool strikeout) {
+int_least32_t font_family::get_line_spacing(intmax_t font_family, int_least32_t em_height, bool bold, bool italic, bool underline, bool strikeout) {
   return 0;
 }
 
-std::string font_family::get_name(intptr_t font_family, int32_t language) {
+std::string font_family::get_name(intmax_t font_family, int_least32_t language) {
   return *reinterpret_cast<std::string*>(font_family);
 }
 
-bool font_family::is_style_available(intptr_t font_family, bool bold, bool italic, bool underline, bool strikeout) {
+bool font_family::is_style_available(intmax_t font_family, bool bold, bool italic, bool underline, bool strikeout) {
   return true;
 }
